Static

+++ pcap/pkt/ip6_unsafe.go
@@ -1,12 +1,14 @@
-// Copyright 2013 The golibpcap Authors. All rights reserved.                      
-// Use of this source code is governed by a BSD-style                              
-// license that can be found in the LICENSE file.                                  
+// Copyright 2013 The golibpcap Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style
+// license that can be found in the LICENSE file.
 
 // +build !safe,!appengine
 
 package pkt
 
 /*
+#include <sys/types.h>
+#include <netinet/in.h>
 #include <netinet/ip6.h>
 */
 import "C"
@@ -36,6 +38,6 @@ func NewIp6Hdr(p unsafe.Pointer) (*Ip6Hdr, unsafe.Pointer) {
 	ip6Hdr.DstAddr = net.IP(C.GoBytes(unsafe.Pointer(&ip6Hdr.cptr.ip6_dst), 16))
 	u := (*C.struct_ip6_hdrctl)(unsafe.Pointer(&ip6Hdr.cptr.ip6_ctlun))
 	ip6Hdr.NextHeader = uint8(u.ip6_un1_nxt)
-	ip6Hdr.PayloadLen = uint16(C.ntohs(u.ip6_un1_plen))
+	ip6Hdr.PayloadLen = uint16(C.ntohs(C.uint16_t(u.ip6_un1_plen)))
 	return ip6Hdr, ip6Hdr.payload
 }
